refactor(provider): move OpenCodeProvider class to provider/OpenCodeProvider.ts
Move the 1100-line provider class out of extension.ts into its own module
(one cohesive concern). Also extract the provider command helpers
(configureUtilityModels / toggleProviderEnabled) into commands/providers.ts
which both the class and activate() import.

extension.ts is now ~630 lines (was 4653) and contains only wiring:
activate (provider registration + command handlers), the agent-window
helpers, model-picker diagnostics and the thinking-effort picker.

Behavior-preserving; compile + 291 tests + test-retry + lint green.
